auth-api/requirements.txt
fastapi
uvicorn[standard]
google-auth
requests